refactor: move shipping method matching cart logic for reuse
We need it in the cart setShippingMethod action
