+++ src/repositories/shipping-method/index.ts
@@ -1,14 +1,12 @@
 import {
-	InvalidOperationError,
 	type ShippingMethod,
 	type ShippingMethodDraft,
 	type ZoneRate,
 	type ZoneRateDraft,
 	type ZoneReference,
 } from "@commercetools/platform-sdk";
-import { CommercetoolsError } from "~src/exceptions";
 import { getBaseResourceProperties } from "../../helpers";
-import { markMatchingShippingRate } from "../../shippingCalculator";
+import { getShippingMethodsMatchingCart } from "../../shipping";
 import { AbstractStorage } from "../../storage/abstract";
 import {
 	AbstractResourceRepository,
@@ -69,57 +67,7 @@ export class ShippingMethodRepository extends AbstractResourceRepository<"shippi
 			return undefined;
 		}
 
-		if (!cart.shippingAddress?.country) {
-			throw new CommercetoolsError<InvalidOperationError>({
-				code: "InvalidOperation",
-				message: `The cart with ID '${cart.id}' does not have a shipping address set.`,
-			});
-		}
-
-		// Get all shipping methods that have a zone that matches the shipping address
-		const zones = this._storage.query<"zone">(context.projectKey, "zone", {
-			where: [`locations(country="${cart.shippingAddress.country}"))`],
-			limit: 100,
-		});
-		const zoneIds = zones.results.map((zone) => zone.id);
-		const shippingMethods = this.query(context, {
-			"where": [
-				`zoneRates(zone(id in (:zoneIds)))`,
-				`zoneRates(shippingRates(price(currencyCode="${cart.totalPrice.currencyCode}")))`,
-			],
-			"var.zoneIds": zoneIds,
-			"expand": params.expand,
-		});
-
-		// Make sure that each shipping method has exactly one shipping rate and
-		// that the shipping rate is marked as matching
-		const results = shippingMethods.results
-			.map((shippingMethod) => {
-				// Iterate through the zoneRates, process the shipping rates and filter
-				// out all zoneRates which have no matching shipping rates left
-				const rates = shippingMethod.zoneRates
-					.map((zoneRate) => ({
-						zone: zoneRate.zone,
-
-						// Iterate through the shippingRates and mark the matching ones
-						// then we filter out the non-matching ones
-						shippingRates: zoneRate.shippingRates
-							.map((rate) => markMatchingShippingRate(cart, rate))
-							.filter((rate) => rate.isMatching),
-					}))
-					.filter((zoneRate) => zoneRate.shippingRates.length > 0);
-
-				return {
-					...shippingMethod,
-					zoneRates: rates,
-				};
-			})
-			.filter((shippingMethod) => shippingMethod.zoneRates.length > 0);
-
-		return {
-			...shippingMethods,
-			results: results,
-		};
+		return getShippingMethodsMatchingCart(context, this._storage, cart, params);
 	}
 
 	private _transformZoneRateDraft(
